backend/pyproject.toml
[project]
name = "blog-backend"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "langchain",
    "langchain-openai",
    "langchain-google-genai",
    "python-dotenv",
    "faiss-cpu",
    "pypdf",
    "unstructured[md]",
    "langchain-community>=0.3.26",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
